amplify-lambda-basic-ops/requirements.txt
git+https://github.com/gaiin-platform/pycommon.git@pre-release/v0.0.1
boto3==1.38.41
httpx==0.28.1
jsonschema==4.24.0
pillow==11.2.1
pydantic==2.11.7
python-dotenv==1.1.0
python-jose==3.5.0
pyyaml==6.0.2
requests==2.32.4
